factor out helper functions to separate files
